Change microstep to be at least 32 bits

+++ org.lflang/src/lib/core/platform/lf_linux_support.h
@@ -60,7 +60,7 @@ typedef long long int _interval_t;
 /**
  * Microstep instant.
  */
-typedef unsigned int _microstep_t;
+typedef unsigned long int _microstep_t;
 
 
 // The underlying physical clock for Linux
